+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.1.0 (unreleased)
 
 - Added method to Journal to create paper_id from issue_id
+- Added method to Journal to get a DOI based on a paper id
 
 ## 0.0.3 (2021-10-08)